add email to supabase on installation and create_user_request
